+++ lib/ruby_challenge/services/segment_parsers/hotel.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RubyChallenge
+  module SegmentParsers
+    class Hotel < Base; end
+  end
+end