+++ spec/services/segment_parsers/hotel_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe RubyChallenge::SegmentParsers::Hotel do
+  include_context "segment parsers default tests"
+end